For auto-activating. Revise_1
Seperate go func() goroutine from StartServer() and invoke
AutoIngestor() from ingestor.go.

Take them as 2 missions when start ingestor. One is to run a new
grouting (go server.AutoIngestor()) behind the scene to activate
capture, and the other is to initiate web server in main goroutine
(server.StartServer()).

Whenever server receives a request and the CaptureFlag is true, which
means the capture is running, just return the response directly.